Move take_over tests to their own test file
